Moved all the files one folder up for consistency
The app is now also built with yarn instead of npm due to an error when building.

Right now I cant test if the app works with an iframe, since the app that is deployed is using pop-ups for authentication instead of redirects, which as far as I know, dont work well there.